+++ src/examples/basicLanguage.ts
@@ -53,6 +53,9 @@ export default function basicLanguage () {
         precedence: [
             ['/', '*'],
             ['-', '+'],
+            // We can use 'pow' as a binary op since it takes exactly two
+            // arguments and has an explicit precedence:
+            ['pow'],
             // We can alter associativity of ops as well (right or left):
             { ops: ['='], associativity: 'right' },
             [';']
@@ -60,17 +63,22 @@ export default function basicLanguage () {
     })
 
     assert.equal(angu.evaluate('"hello " + "world"', ctx()).value, "hello world")
+
     // We can pass in local variables at eval time:
     assert.equal(angu.evaluate('"hello " + w', ctx(), { w: 'world' }).value, "hello world")
+
     assert.equal(angu.evaluate('"hello" * 4', ctx()).value, "hellohellohellohello")
     assert.equal(angu.evaluate('1 + 2 + 3', ctx()).value, 6)
     assert.equal(angu.evaluate('1 + 2 + 3 / 3', ctx()).value, 4)
     assert.equal(angu.evaluate('pow(1 + 2 +  3/3, 2) / 2', ctx()).value, 8)
+
     // We can pass in local functions at eval time, which override those in scope:
     assert.equal(angu.evaluate('pow(1 + 2, 2)', ctx(), { pow: (a: Any, b: Any) => a.eval() + b.eval() }).value, 5)
-    // Functions can be used inline using back ticks:
-    assert.equal(angu.evaluate("2 `pow` 3", ctx()).value, 8)
-    assert.equal(angu.evaluate("2 `pow` 3", ctx(), { pow: (a: Any, b: Any) => a.eval() + b.eval() }).value, 5)
+
+    // String functions can be used inline IF they take 2 args *and* have an explicit precedence:
+    assert.equal(angu.evaluate("2 pow 3", ctx()).value, 8)
+    assert.equal(angu.evaluate("2 pow 3", ctx(), { pow: (a: Any, b: Any) => a.eval() + b.eval() }).value, 5)
+
     assert.equal(angu.evaluate(' log10(100)  +2 -2', ctx()).value, 2)
     assert.equal(angu.evaluate('foo = 8; foo = 10; bar = 2; foo * bar', ctx()).value, 20)